Pages for add class
Added pages for adding values to a new class, requires linking to
memory, from memory, listview will show the list of classes.

To get it in memory, a .java will need to be made that adds all the
class info to a class object, and saved to device memory in a list of
classes.

This will be the referenced list of classes when printed, and when
setting up a notification.